Separated RequsteBodies and Data Objects
And improved Controllers

+++ deliveryService/src/main/java/edu/tum/ase/deliveryService/controller/DeliveryController.java
@@ -1,11 +1,11 @@
 package edu.tum.ase.deliveryService.controller;
 
-import edu.tum.ase.deliveryService.exceptions.BoxHasActiveDeliveriesException;
-import edu.tum.ase.deliveryService.exceptions.BoxHasDeliveredDeliveriesException;
-import edu.tum.ase.deliveryService.exceptions.DeliveryNotPartOfBoxException;
+import edu.tum.ase.deliveryService.exceptions.DeliveryStatusException;
+import edu.tum.ase.deliveryService.exceptions.SingleCustomerPerBoxViolationException;
 import edu.tum.ase.deliveryService.model.Box;
 import edu.tum.ase.deliveryService.model.Delivery;
 import edu.tum.ase.deliveryService.model.DeliveryStatus;
+import edu.tum.ase.deliveryService.request.DeliveryRequest;
 import edu.tum.ase.deliveryService.service.BoxService;
 import edu.tum.ase.deliveryService.service.DeliveryService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,7 +13,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Collection;
+import javax.validation.Valid;
 import java.util.List;
 
 @RestController
@@ -26,35 +26,77 @@ public class DeliveryController {
     @Autowired
     BoxService boxService;
 
+
+    //##################################################################################################################
+    // GET mappings
+
     @GetMapping("")
     @PreAuthorize("hasRole('DISPATCHER')")
     public List<Delivery> getAllDeliveries() {
         return deliveryService.getAllDeliveries();
     }
 
-    @GetMapping("{id}")
+    @GetMapping("{deliveryId}")
     @PreAuthorize("hasRole('DISPATCHER')")
-    public Delivery getDelivery(@PathVariable String id) {
-        return deliveryService.findById(id);
+    public Delivery getDelivery(@PathVariable String deliveryId) {
+        return deliveryService.findById(deliveryId);
     }
+
+    //##################################################################################################################
+    // POST mappings
 
     @PostMapping("{boxId}")
     @PreAuthorize("hasRole('DISPATCHER')")
-    public Box createAndAddDelivery(@RequestBody Delivery delivery, @PathVariable String boxId) {
+    public Delivery createAndAddDelivery(@Valid @RequestBody DeliveryRequest deliveryRequest, @PathVariable String boxId) {
+        Delivery delivery = new Delivery();
+        deliveryRequest.apply(delivery);
+
         Box box = boxService.findById(boxId);
         box.addDelivery(delivery);
+        boxService.updateBox(box);
 
-        return boxService.updateBox(box);
+        return delivery; // delivery is persisted in line above
     }
 
-    @PutMapping("{id}")
+    //##################################################################################################################
+    // PUT mappings
+
+    @PutMapping("{deliveryId}")
     @PreAuthorize("hasRole('DISPATCHER')")
-    public Delivery updateDelivery(@RequestBody Delivery newdelivery, @PathVariable String id) {
-        newdelivery.setId(id);
-        return deliveryService.updateDelivery(newdelivery);
+    public Delivery updateDelivery(@Valid @RequestBody DeliveryRequest deliveryRequest, @PathVariable String deliveryId) {
+        Delivery delivery = deliveryService.findById(deliveryId);
+        deliveryRequest.apply(delivery);
+
+        if (delivery.getStatus() == DeliveryStatus.DELIVERED)
+            boxService.clearDeliveryAssignment(delivery);
+
+        return deliveryService.updateDelivery(delivery); // delivery is persisted in line above
     }
 
-    @PutMapping("{boxId}/{id}")
+    @PutMapping("{deliveryId}/assign/{boxId}")
+    @PreAuthorize("hasRole('DISPATCHER')")
+    public Delivery assignDelivery(@PathVariable String deliveryId, @PathVariable String boxId) {
+        Delivery delivery = deliveryService.findById(deliveryId);
+        Box box = boxService.findById(boxId);
+
+        if (!delivery.getStatus().canBeReassigned()) {
+            throw new DeliveryStatusException();
+        }
+
+        if (!box.getDeliveries().get(0).getCustomer().equals(delivery.getCustomer())) {
+            throw new SingleCustomerPerBoxViolationException();
+        }
+
+        boxService.clearDeliveryAssignment(delivery);
+        box.addDelivery(delivery);
+
+        boxService.updateBox(box);
+        return delivery; // delivery is persisted in line above
+    }
+
+
+    // I don't think this is necessary
+    /*@PutMapping("{boxId}/{id}")
     @PreAuthorize("hasRole('DISPATCHER')")
     public Box removeDeliveryFromBox(@PathVariable String boxId, @PathVariable String id) {
         Box box = boxService.findById(boxId);
@@ -70,19 +112,22 @@ public class DeliveryController {
             }
         }
         throw new DeliveryNotPartOfBoxException();
-    }
+    }*/
+
+
+    //##################################################################################################################
+    // DELETE mappings
 
     @DeleteMapping("{id}")
     @PreAuthorize("hasRole('DISPATCHER')")
     public HttpStatus deleteDelivery(@PathVariable String id) {
         Delivery delivery = deliveryService.findById(id);
 
-        //TODO: Check if delivery is still used in Box --> if yes, delete is not allowed!
+        if (!delivery.getStatus().canBeRemoved()) {
+            throw new DeliveryStatusException();
+        }
 
         deliveryService.deleteDelivery(delivery);
         return HttpStatus.OK;
     }
-
-
-    // TODO: Maybe GET BOX of delivery?
 }
